+++ NEWS.md
@@ -1,6 +1,8 @@
 # corrr (development version)
 
-# corrr 0.3.2.9002
+# corrr 0.3.2.9003
+
+- Adds `remove.dups` argument to `stretch()`.  It removes duplicates with out removing all NAs (#57)
 
 - Adds `dice()` function, wraps `focus(x,..., mirror = TRUE)` (#64)